+++ app/controllers/plans_controller.rb
@@ -2,12 +2,12 @@
 
 class PlansController < ApplicationController
   def new
-    plan = Plan.create(category: "catering")
+    plan = Plan.create(category: "catering", next_entry_id: ENV["CONTENTFUL_PLANNING_START_ENTRY_ID"])
     redirect_to new_plan_question_path(plan)
   end
 
   def show
-    @plan = Plan.find(plan_id)
+    @plan = Plan.includes(questions: [:answer]).find(plan_id)
   end
 
   private