+++ app/controllers/questions_controller.rb
@@ -2,19 +2,20 @@
 
 class QuestionsController < ApplicationController
   rescue_from GetContentfulEntry::EntryNotFound do |exception|
-    @exception = exception
     render "errors/contentful_entry_not_found", status: 500
+  end
+
+  rescue_from CreatePlanningQuestion::UnexpectedContentType do |exception|
+    render "errors/unexpected_contentful_type", status: 500
   end
 
   def new
     @plan = Plan.find(plan_id)
+
+    redirect_to plan_path(@plan) unless @plan.next_entry_id.present?
+
     @question = CreatePlanningQuestion.new(plan: @plan).call
     @answer = Answer.new
-    # TODO: Creating a question requires us to check externally if one exists
-    # based on the previous question. Instead of looping through at the start,
-    # we assume there is and try to create one. If not, we jump to the end.
-
-    redirect_to plan_path(@plan) if @question.nil?
   end
 
   private